% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_color_warmCrush_d}
\alias{scale_color_warmCrush_d}
\title{Colors for discrete value in a plot.}
\usage{
scale_color_warmCrush_d(primary = "Green", other = "Copper", direction = 1)
}
\arguments{
\item{primary}{This is the primary color chosen for plotting.}

\item{other}{This other color for the palette.}

\item{direction}{This is a numeric value to set the direction of the color scale i.e. if the direction is < 0, the direction of colour changes.}
}
\value{
Color outline for the plots
}
\description{
Colors for discrete value in a plot.
}
