% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_color_warmCrush_seq_c}
\alias{scale_color_warmCrush_seq_c}
\title{Colors for continuous scale}
\usage{
scale_color_warmCrush_seq_c(name = "Rust", direction = 1, amount = 0.7)
}
\arguments{
\item{name}{Name of the color which is default.}

\item{direction}{This is a numeric value to set the direction of the color scale i.e. if the direction is < 0, the direction of colour changes.}

\item{amount}{This is the level of luminance.A negative value changes the direction of luminance.}
}
\value{
Color in the plot with different levels of luminance.
}
\description{
Colors for continuous scale
}
