% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_colour_warmCrush_d}
\alias{scale_colour_warmCrush_d}
\title{this is title}
\usage{
scale_colour_warmCrush_d(primary = "Green", other = "Copper", direction = 1)
}
\arguments{
\item{direction}{a numeric value to set the direction of the color scale if direction is < 0 switch the direction}
}
\value{
colour
}
\description{
this is title
}
