% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_warmCrush_seq_c}
\alias{scale_colour_warmCrush_seq_c}
\title{Title}
\usage{
scale_colour_warmCrush_seq_c(name = "Rust", direction = 1, amount = 0.7)
}
\arguments{
\item{direction}{a numeric value to set the direction of the color scale if direction is < 0 switch the direction}

\item{amount}{scales of the colour}
}
\value{
colour
}
\description{
Title
}
