% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_fill_warmCrush_d}
\alias{scale_fill_warmCrush_d}
\title{Fill for discrete value in a plot.}
\usage{
scale_fill_warmCrush_d(primary = "Green", other = "Copper", direction = 1)
}
\arguments{
\item{primary}{This is the primary colour chosen for plotting.}

\item{other}{This other colour for the palette.}

\item{direction}{This is a numeric value to set the direction of the color scale i.e. if the direction is < 0, the direction of colour changes.}
}
\value{
colours inside the plot
}
\description{
Fill for discrete value in a plot.
}
