% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{warmCrush_colours}
\alias{warmCrush_colours}
\title{warmCrush colours}
\format{
An object of class \code{list} of length 8.
}
\usage{
warmCrush_colours
}
\description{
warmCrush colours
}
\examples{
warmCrush_colours
}
\keyword{datasets}
