% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{warmCrush_pal}
\alias{warmCrush_pal}
\title{this is title}
\usage{
warmCrush_pal(primary = "Green", other = "Copper", direction = 1)
}
\arguments{
\item{direction}{a numeric value to set the direction of the color scale if direction is < 0 switch the direction}
}
\value{
colour
}
\description{
this is title
}
