% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{warmCrush_colours}
\alias{warmCrush_colours}
\title{urbanC Color Palette:

\code{warmCrush_pal}}
\format{
An object of class \code{list} of length 8.
}
\usage{
warmCrush_colours
}
\arguments{
\item{direction:}{a numeric value to set the direction of the color scale.If direction is < 0 the color list is reversed.}
}
\description{
urbanC Color Palette:

\code{warmCrush_pal}
}
\details{
Default colors are generated using this warmCrush palette.
}
\section{Arguments}{
NA
}

\section{primary the baseline color for continuous scale}{
NA
}

\keyword{datasets}
