% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{warmCrush_pal}
\alias{warmCrush_pal}
\title{urbanC colour palette warmCrush}
\usage{
warmCrush_pal(primary = "Green", other = "Copper", direction = 1)
}
\arguments{
\item{primary}{This is the primary colour chosen for plotting.}

\item{other}{This other colour for the palette.}

\item{direction}{This is a numeric value to set the direction of the color scale i.e. if the direction is < 0, the direction of colour changes.}
}
\value{
colours for the palette
}
\description{
urbanC colour palette warmCrush
}
